Updating to Firefox version 98.0.2

+++ xpcom/string/nsString.h
@@ -52,7 +52,7 @@ class NS_LossyConvertUTF16toASCII : public nsAutoCString {
     LossyAppendUTF16toASCII(mozilla::MakeStringSpan(aString), *this);
   }
 
-  NS_LossyConvertUTF16toASCII(const char16ptr_t aString, uint32_t aLength) {
+  NS_LossyConvertUTF16toASCII(const char16ptr_t aString, size_t aLength) {
     LossyAppendUTF16toASCII(
         Substring(static_cast<const char16_t*>(aString), aLength), *this);
   }
@@ -74,7 +74,7 @@ class NS_ConvertASCIItoUTF16 : public nsAutoString {
     AppendASCIItoUTF16(mozilla::MakeStringSpan(aCString), *this);
   }
 
-  NS_ConvertASCIItoUTF16(const char* aCString, uint32_t aLength) {
+  NS_ConvertASCIItoUTF16(const char* aCString, size_t aLength) {
     AppendASCIItoUTF16(Substring(aCString, aLength), *this);
   }
 
@@ -102,7 +102,7 @@ class NS_ConvertUTF16toUTF8 : public nsAutoCString {
     AppendUTF16toUTF8(mozilla::MakeStringSpan(aString), *this);
   }
 
-  NS_ConvertUTF16toUTF8(const char16ptr_t aString, uint32_t aLength) {
+  NS_ConvertUTF16toUTF8(const char16ptr_t aString, size_t aLength) {
     AppendUTF16toUTF8(Substring(static_cast<const char16_t*>(aString), aLength),
                       *this);
   }
@@ -128,7 +128,7 @@ class NS_ConvertUTF8toUTF16 : public nsAutoString {
     AppendUTF8toUTF16(mozilla::MakeStringSpan(aCString), *this);
   }
 
-  NS_ConvertUTF8toUTF16(const char* aCString, uint32_t aLength) {
+  NS_ConvertUTF8toUTF16(const char* aCString, size_t aLength) {
     AppendUTF8toUTF16(Substring(aCString, aLength), *this);
   }
